+++ NEWS.md
@@ -16,6 +16,7 @@ Every function that adds a serie to chart (e.g.: `e_bar`) have been turned into 
 - Fixed a few issues with `e_mark*` family, labels and multiple marks correctly work.
 - Initialisation function now take `reorder` argument to specify whether to reorder numeric values. [#126](https://github.com/JohnCoene/echarts4r/issues/126)
 - Added convenience `e_hide_grid_lines` to easily hide grid lines.
+- Added `e_register_theme` to register a global theme, see [#127](https://github.com/JohnCoene/echarts4r/issues/127).
 
 # echarts4r 0.2.3